pyproject.toml
[project]
name = "multi-agent-beads"
version = "0.1.0"
description = "Multi-agent SDLC orchestration system with concurrent Developer, QA, Tech Lead, Manager, and Code Reviewer agents"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "pyyaml>=6.0",
    "click>=8.1.0",
]

[project.scripts]
mab = "mab.cli:main"
bd-close = "mab.bd_close:main"

[project.optional-dependencies]
dev = [
    "httpx>=0.26.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-playwright>=0.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "types-PyYAML>=6.0.0",
]

[tool.uv]
dev-dependencies = [
    "httpx>=0.26.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-playwright>=0.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "types-PyYAML>=6.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.hatch.build.targets.wheel]
packages = ["dashboard", "scripts", "mab"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
